+++ LossFunction.hpp
@@ -26,9 +26,9 @@ public:
 		const Eigen::VectorXd& targets) const override {
 		assert(predictions.size() == targets.size());
 
-		int classNum = predictions.size();
+		size_t classNum = predictions.size();
 		double predictLoss = 0.0;
-		for (int c = 0; c < classNum; ++c) {
+		for (size_t c = 0; c < classNum; ++c) {
 			double error = predictions[c] - targets[c];
 			predictLoss += std::pow(error, 2.0);
 		}
@@ -55,10 +55,10 @@ public:
 		const Eigen::VectorXd& targets) const override {
 
 		assert(predictions.size() == targets.size());
-		int classNum = predictions.size();
+		size_t classNum = predictions.size();
 
 		double loss = 0.0;
-		for (int c = 0; c < classNum; ++c) {
+		for (size_t c = 0; c < classNum; ++c) {
 
 			loss -= targets(c) * std::log(std::max(predictions(c), _epsilon));
 		}
@@ -71,10 +71,10 @@ public:
 
 		assert(predictions.size() == targets.size());
 
-		int classNum = predictions.size();
+		size_t classNum = predictions.size();
 		Eigen::VectorXd gradientCE(classNum);
 
-		for (int c = 0; c < classNum; ++c) {
+		for (size_t c = 0; c < classNum; ++c) {
 
 			gradientCE(c) = -targets(c) / (predictions(c) + _epsilon);
 		}
